requirements.txt
pymysql
peewee
websocket-client
openpyxl
pandas